+++ CHANGELOG.md
@@ -21,6 +21,7 @@ PyCheval adheres to [semantic versioning](https://semver.org/).
 - Validate that the seller has a tax registration number.
 - Validate that the tax amounts match.
 - Validate that each line charge has a reason code or reason text.
+- `ProductClassification` now requires a `listID` attribute.
 
 ## [0.1.0] - 2025-05-07